# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pitlane Automotive Group API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface MechanicLoginInput {
  code: string;
  password: string;
}

export interface AuthResult {
  success: boolean;
  role: string;
}

export interface Mechanic {
  id: number;
  name: string;
  /** @nullable */
  contact?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  location?: string | null;
  code: string;
  discount: number;
  revShare: number;
}

export interface MechanicAuthResult {
  success: boolean;
  mechanic: Mechanic;
}

export type SubscriberTier =
  (typeof SubscriberTier)[keyof typeof SubscriberTier];

export const SubscriberTier = {
  basic: "basic",
  standard: "standard",
  annual: "annual",
} as const;

export type SubscriberConsent =
  (typeof SubscriberConsent)[keyof typeof SubscriberConsent];

export const SubscriberConsent = {
  yes: "yes",
  pending: "pending",
} as const;

export interface Subscriber {
  id: number;
  fname: string;
  lname: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  suburb?: string | null;
  /** @nullable */
  make?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  year?: string | null;
  tier: SubscriberTier;
  /** @nullable */
  mechId?: number | null;
  /** @nullable */
  nextService?: string | null;
  consent: SubscriberConsent;
  joinDate: string;
}

export type SubscriberInputTier =
  (typeof SubscriberInputTier)[keyof typeof SubscriberInputTier];

export const SubscriberInputTier = {
  basic: "basic",
  standard: "standard",
  annual: "annual",
} as const;

export type SubscriberInputConsent =
  (typeof SubscriberInputConsent)[keyof typeof SubscriberInputConsent];

export const SubscriberInputConsent = {
  yes: "yes",
  pending: "pending",
} as const;

export interface SubscriberInput {
  fname: string;
  lname: string;
  phone?: string;
  email?: string;
  suburb?: string;
  make?: string;
  model?: string;
  year?: string;
  tier: SubscriberInputTier;
  /** @nullable */
  mechId?: number | null;
  /** @nullable */
  nextService?: string | null;
  consent?: SubscriberInputConsent;
}

export type SubscriberUpdateTier =
  (typeof SubscriberUpdateTier)[keyof typeof SubscriberUpdateTier];

export const SubscriberUpdateTier = {
  basic: "basic",
  standard: "standard",
  annual: "annual",
} as const;

export type SubscriberUpdateConsent =
  (typeof SubscriberUpdateConsent)[keyof typeof SubscriberUpdateConsent];

export const SubscriberUpdateConsent = {
  yes: "yes",
  pending: "pending",
} as const;

export interface SubscriberUpdate {
  fname?: string;
  lname?: string;
  phone?: string;
  email?: string;
  suburb?: string;
  make?: string;
  model?: string;
  year?: string;
  tier?: SubscriberUpdateTier;
  /** @nullable */
  mechId?: number | null;
  /** @nullable */
  nextService?: string | null;
  consent?: SubscriberUpdateConsent;
}

export interface MechanicInput {
  name: string;
  contact?: string;
  phone?: string;
  email?: string;
  location?: string;
  discount?: number;
  revShare?: number;
  password?: string;
}

export interface MechanicUpdate {
  name?: string;
  contact?: string;
  phone?: string;
  email?: string;
  location?: string;
  discount?: number;
  revShare?: number;
  password?: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
} as const;

export interface Booking {
  id: number;
  subId: number;
  mechId: number;
  service: string;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  value?: number | null;
  status: BookingStatus;
}

export type BookingInputStatus =
  (typeof BookingInputStatus)[keyof typeof BookingInputStatus];

export const BookingInputStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
} as const;

export interface BookingInput {
  subId: number;
  mechId: number;
  service: string;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  value?: number | null;
  status: BookingInputStatus;
}

export type BookingUpdateStatus =
  (typeof BookingUpdateStatus)[keyof typeof BookingUpdateStatus];

export const BookingUpdateStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
} as const;

export interface BookingUpdate {
  subId?: number;
  mechId?: number;
  service?: string;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  value?: number | null;
  status?: BookingUpdateStatus;
}

export type DashboardStatsTierBreakdown = {
  basic: number;
  standard: number;
  annual: number;
};

export type DashboardStatsRevenueBreakdown = {
  basicAnnual: number;
  standardAnnual: number;
  annualTierAnnual: number;
  grossAnnual: number;
  pitlaneNet: number;
  weeklyRunRate: number;
};

export type ServiceAlertAlertStatus =
  (typeof ServiceAlertAlertStatus)[keyof typeof ServiceAlertAlertStatus];

export const ServiceAlertAlertStatus = {
  overdue: "overdue",
  upcoming: "upcoming",
} as const;

export interface ServiceAlert {
  subscriberId: number;
  subscriberName: string;
  vehicleDescription?: string;
  tier: string;
  nextService: string;
  alertStatus: ServiceAlertAlertStatus;
  daysUntil: number;
  workshopName?: string;
}

export interface DashboardStats {
  totalSubscribers: number;
  activeMechanics: number;
  totalBookings: number;
  totalRevenue: number;
  weeklyRunRate: number;
  recentSubscribers: Subscriber[];
  upcomingBookings: Booking[];
  tierBreakdown: DashboardStatsTierBreakdown;
  revenueBreakdown: DashboardStatsRevenueBreakdown;
  serviceAlerts: ServiceAlert[];
}

export interface MechanicStats {
  totalSubscribers: number;
  totalBookings: number;
  revenueEarned: number;
  upcomingBookings: Booking[];
  subscribers: Subscriber[];
}

export type SignupInputTier =
  (typeof SignupInputTier)[keyof typeof SignupInputTier];

export const SignupInputTier = {
  basic: "basic",
  standard: "standard",
  annual: "annual",
} as const;

export interface SignupInput {
  fname: string;
  lname: string;
  phone?: string;
  email?: string;
  suburb?: string;
  make?: string;
  model?: string;
  year?: string;
  tier: SignupInputTier;
}