# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pitlane Automotive Group API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  subId: number;
  mechId: number;
  service: string;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  value?: number | null;
  status: BookingStatus;
}