# lib/api-zod/src/generated/types/mechanicAuthResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pitlane Automotive Group API
 * OpenAPI spec version: 0.1.0
 */
import type { Mechanic } from "./mechanic";

export interface MechanicAuthResult {
  success: boolean;
  mechanic: Mechanic;
}