# lib/api-zod/src/generated/types/mechanicLoginInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pitlane Automotive Group API
 * OpenAPI spec version: 0.1.0
 */

export interface MechanicLoginInput {
  code: string;
  password: string;
}