# lib/api-zod/src/generated/types/mechanicStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pitlane Automotive Group API
 * OpenAPI spec version: 0.1.0
 */
import type { Booking } from "./booking";
import type { Subscriber } from "./subscriber";

export interface MechanicStats {
  totalSubscribers: number;
  totalBookings: number;
  revenueEarned: number;
  upcomingBookings: Booking[];
  subscribers: Subscriber[];
}